ASSIGNMENT-03/studybuddy/pyproject.toml
[project]
name = "studybuddy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "MUDASSIR ABRAR", email = "161126723+MudassarAbrar@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.15",
    "python-dotenv>=1.1.0",
]

[project.scripts]
studybuddy = "studybuddy.my_study_buddy:my_first_agent"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
